+++ FarmaciaERP/Backend/src/main/java/FarmaciaERP/Application/UseCases/EliminarUsuarioUseCase.java
@@ -11,7 +11,7 @@ public class EliminarUsuarioUseCase {
         this.usuarioRepository = usuarioRepository;
     }
 
-    public void ejecutar(int id) {
+    public void ejecutar(Long id) {
         if (!usuarioRepository.existsById(id)) {
             throw new RuntimeException("El paciente con ID " + id + " no existe.");
         }
